requirements.txt
numpy
opencv-python
onnxruntime-gpu
scikit-learn
scikit-image
tqdm
mxnet
insightface